Crate common_runtime

Modules

adapter

locked_balance

weights

Structs

AdjustmentVariable

The adjustment variable of the runtime. Higher values will cause TargetBlockFullness to change the fees more rapidly.

AnnouncementDepositBase

AnnouncementDepositFactor

AssetLocation

BasicDeposit

BasiliskExtrinsicBaseWeight

Basilisk base weight of an extrinsic This includes weight for payment in non-native currency.

BlockHashCount

BlockLength

Maximum length of block. Up to 5MB.

Burn

CandidacyBond

CooloffPeriod

CouncilMaxMembers

CouncilMaxProposals

CouncilMotionDuration

DataDepositPerByte

DepositBase

DepositFactor

DesiredMembers

DesiredRunnersUp

DiscountedFee

DustingReward

ElectionsPhragmenPalletId

EnactmentPeriod

ExchangeFee

FastTrackVotingPeriod

FieldDeposit

InstantAllowed

LBPExchangeFee

LMPalletId

LaunchPeriod

MaxAdditionalFields

MaxApprovals

MaxAuthorities

MaxCandidates

MaxInRatio

MaxInvulnerables

MaxLocks

MaxOutRatio

MaxPending

MaxProposals

MaxProxies

MaxRegistrars

MaxReserves

MaxSignatories

MaxSubAccounts

MaxVestingSchedules

MaxVotes

MaximumReasonLength

MinCandidates

MinPlannedYieldingPeriods

MinPoolLiquidity

MinTotalFarmRewards

MinTradingLimit

MinVestedTransfer

MinimumDeposit

MinimumMultiplier

Minimum amount of the multiplier. This value cannot be too low. A test case should ensure that combined with AdjustmentVariable, we can recover from the minimum.

MinimumPeriod

MultiPaymentCurrencySetFee

NativeAssetId

NativeExistentialDeposit

NftClass

Offset

Period

PotId

PreimageByteDeposit

ProposalBond

ProposalBondMaximum

ProposalBondMinimum

ProxyDepositBase

ProxyDepositFactor

RegistryStrLimit

SS58Prefix

SpendPeriod

SubAccountDeposit

TargetBlockFullness

The portion of the NORMAL_DISPATCH_RATIO that we adjust the fees with. Blocks filled less than this will decrease the weight and more will increase.

TechnicalMaxMembers

TechnicalMaxProposals

TechnicalMotionDuration

TermDuration

TipCountdown

TipFindersFee

TipReportDepositBase

TipReportDepositPerByte

TransactionByteFee

TreasuryPalletId

UncheckedExtrinsic

Opaque, encoded, unchecked extrinsic. Simple blob to hold an extrinsic without committing to its format and ensure it is serialized correctly.

UncleGenerations

VestingPalletId

VotingBondBase

VotingBondFactor

VotingPeriod

Enums

ProxyType

The type used to represent the kinds of proxying allowed.

Constants

AVERAGE_ON_INITIALIZE_RATIO

We assume that an on-initialize consumes 2.5% of the weight on average, hence a single extrinsic will not be allowed to consume more than AvailableBlockRatio - 2.5%.

CENTS

CORE_ASSET_ID

Core asset id

DAYS

DISCOUNTED_FEE

Discounted XYK fee

DOLLARS

EPOCH_DURATION_IN_BLOCKS

FORTUNE

HOURS

MAXIMUM_BLOCK_WEIGHT

We allow for

MAX_IN_RATIO

Max fraction of pool to sell in single transaction

MAX_OUT_RATIO

Max fraction of pool to buy in single transaction

MILLICENTS

MILLISECS_PER_BLOCK

Since BABE is probabilistic this is the average expected block time that we are targeting. Blocks will be produced at a minimum duration defined by SLOT_DURATION, but some slots will not be allocated to any authority and hence no block will be produced. We expect to have this block time on average following the defined slot duration and the value of c configured for BABE (where 1 - c represents the probability of a slot being empty). This value is only used indirectly to define the unit constants below that are expressed in blocks. The rest of the code should use SLOT_DURATION instead (like the Timestamp pallet for calculating the minimum period).

MINUTES

MIN_POOL_LIQUIDITY

Minimum pool liquidity

MIN_TRADING_LIMIT

Trading limit

NATIVE_EXISTENTIAL_DEPOSIT

NORMAL_DISPATCH_RATIO

We allow Normal extrinsics to fill up the block up to 75%, the rest can be used by Operational extrinsics.

RELAY_CHAIN_ASSET_LOCATION

SECS_PER_BLOCK

SLOT_DURATION

UNITS

WEIGHT_PER_SECOND

Functions

deposit

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex

The type for looking up accounts. We don’t expect more than 4 billion of them, but you never know…

Amount

Signed version of Balance

AssetId

Type for storing the id of an asset.

Balance

Type for storing the balance of an account.

Block

Block type.

BlockNumber

An index to a block.

Hash

A hash of some data used by the chain.

Header

Header type.

Index

Index of a transaction in the chain.

Multiplier

Fee multiplier.

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

Weight

Numeric range of a transaction weight.