Crate common_runtime

Modules

constants

extensions

fee

weights

Expose the auto generated weight files.

Macros

prod_or_testnet_or_local

Macro to set a value (e.g. when using the parameter_types macro) to either a production value or to an environment variable or testing value (in case the frequency-rococo-local feature is selected or in instant-sealing mode). Note that the environment variable is evaluated at compile time.