Crate common_runtime

Re-exports

pub use crate::runtime::Builder;

pub use crate::runtime::Runtime;

Modules

error

global 🔒

Global runtimes

metrics 🔒

Runtime metrics

repeated_task 🔒

runtime

Structs

JoinError

Task failed to execute to completion.

JoinHandle

An owned permission to join on a task (await its termination).

RepeatedTask

Traits

TaskFunction

Task to execute repeatedly.

Functions

bg_runtime

Returns the global bg thread pool.

block_on_bg

Run a future to complete in bg thread pool.

block_on_read

Run a future to complete in read thread pool.

block_on_write

Run a future to complete in write thread pool.

create_runtime

init_global_runtimes

Initialize the global runtimes

read_runtime

Returns the global read thread pool.

spawn_bg

Spawn a future and execute it in bg thread pool.

spawn_blocking_bg

Run the blocking operation in bg thread pool.

spawn_blocking_read

Run the blocking operation in read thread pool.

spawn_blocking_write

Run the blocking operation in write thread pool.

spawn_read

Spawn a future and execute it in read thread pool.

spawn_write

Spawn a future and execute it in write thread pool.

write_runtime

Returns the global write thread pool.

Type Definitions

BoxedTaskFunction