Module common_runtime::constants

Modules

currency

Structs

Burn

How much of the treasury to burn, if funds remain at the end of the SpendPeriod Set to zero until the economic system is setup and stabilized

CollatorKickThreshold

CollatorMaxCandidates

CollatorMaxInvulnerables

CollatorMinCandidates

CooloffPeriod

CouncilMotionDuration

EnactmentPeriod

FastTrackVotingPeriod

LaunchPeriod

MaxApprovals

Maximum number of approved proposals per Spending Period Set to 64 or 16 per week

MaxDataSize

Clone + Debug + Eq implementation for u32 types

MessagesMaxPayloadSizeBytes

MinVestedTransfer

MinimumDeposit

MinimumPeriod

NeverDepositIntoId

PreimageBaseDeposit

PreimageByteDeposit

ProposalBondMaximum

Minimum bond for a treasury proposal

ProposalBondMinimum

Minimum bond for a treasury proposal

ProposalBondPercent

Bond amount a treasury request must put up to make the proposal This will be transferred to OnSlash if the proposal is rejected

SchemasMaxRegistrations

SpendPeriod

TCMotionDuration

TransactionByteFee

Relay Chain TransactionByteFee / 10

TreasuryPalletId

Keyless account that holds the money for the treasury

VestingPalletId

VotingPeriod

Constants

AVERAGE_ON_INITIALIZE_RATIO

We assume that ~5% of the block weight is consumed by on_initialize handlers. This is used to limit the maximal weight of a single extrinsic.

DAYS

FREQUENCY_LOCAL_TOKEN

FREQUENCY_ROCOCO_TOKEN

FREQUENCY_TOKEN

HOURS

MAXIMUM_BLOCK_WEIGHT

We allow for 0.5 of a second of compute with a 12 second average block time.

MILLISECS_PER_BLOCK

This determines the average expected block time that we are targeting. Blocks will be produced at a minimum duration defined by SLOT_DURATION. SLOT_DURATION is picked up by pallet_timestamp which is in turn picked up by pallet_aura to implement fn slot_duration().

MINUTES

NORMAL_DISPATCH_RATIO

We allow Normal extrinsics to fill up the block up to 75%, the rest can be used by Operational extrinsics.

ORML_MAX_VESTING_SCHEDULES

SLOT_DURATION

TOKEN_DECIMALS

TREASURY_PALLET_ID

Generates the pallet “account” 5EYCAe5ijiYfyeZ2JJCGq56LmPyNRAKzpG4QkoQkkQNB5e6Z

Type Definitions

AuraMaxAuthorities

AuthorshipUncleGenerations

BalancesMaxLocks

BalancesMaxReserves

CouncilMaxProposals

DemocracyMaxProposals

DemocracyMaxVotes

FIFTY

FrameSystemMaxConsumers

HUNDRED

MessagesMaxPerBlock

MsaMaxKeys

MsaMaxProviderNameSize

PreimageMaxSize

Preimage maximum size set to 4 MB Expected to be removed in Polkadot v0.9.31

SchedulerMaxScheduledPerBlock

SchemasMaxBytesBoundedVecLimit

SchemasMinModelSizeBytes

SessionOffset

SessionPeriod

TCMaxMembers

TCMaxProposals

TransactionPaymentOperationalFeeMultiplier

ZERO