Module common_runtime::global

Global runtimes

Macros

define_global_runtime_spawn 🔒

define_spawn 🔒

Structs

ConfigRuntimes 🔒

GlobalRuntimes 🔒

RuntimeOptions

The options for the global runtimes.

Constants

BG_WORKERS 🔒

HB_WORKERS 🔒

READ_WORKERS 🔒

WRITE_WORKERS 🔒

Statics

CONFIG_RUNTIMES 🔒

GLOBAL_RUNTIMES 🔒

Functions

bg_runtime

Returns the global bg thread pool.

block_on_bg

Run a future to complete in bg thread pool.

block_on_hb

Run a future to complete in hb thread pool.

block_on_read

Run a future to complete in read thread pool.

block_on_write

Run a future to complete in write thread pool.

create_runtime

hb_runtime

Returns the global hb thread pool.

init_global_runtimes

Initialize the global runtimes

read_runtime

Returns the global read thread pool.

spawn_bg

Spawn a future and execute it in bg thread pool.

spawn_blocking_bg

Run the blocking operation in bg thread pool.

spawn_blocking_hb

Run the blocking operation in hb thread pool.

spawn_blocking_read

Run the blocking operation in read thread pool.

spawn_blocking_write

Run the blocking operation in write thread pool.

spawn_hb

Spawn a future and execute it in hb thread pool.

spawn_read

Spawn a future and execute it in read thread pool.

spawn_write

Spawn a future and execute it in write thread pool.

write_runtime

Returns the global write thread pool.