common_runtime

Module global

Global runtimes

Macros

define_global_runtime_spawn 🔒

define_spawn 🔒

Structs

ConfigRuntimes 🔒

GlobalRuntimes 🔒

RuntimeOptions

The options for the global runtimes.

Constants

COMPACT_WORKERS 🔒

GLOBAL_WORKERS 🔒

HB_WORKERS 🔒

Statics

CONFIG_RUNTIMES 🔒

GLOBAL_RUNTIMES 🔒

Functions

block_on_compact

Run a future to complete in compact thread pool.

block_on_global

Run a future to complete in global thread pool.

block_on_hb

Run a future to complete in hb thread pool.

compact_runtime

Returns the global compact thread pool.

create_runtime

global_runtime

Returns the global global thread pool.

hb_runtime

Returns the global hb thread pool.

init_global_runtimes

Initialize the global runtimes

spawn_blocking_compact

Run the blocking operation in compact thread pool.

spawn_blocking_global

Run the blocking operation in global thread pool.

spawn_blocking_hb

Run the blocking operation in hb thread pool.

spawn_compact

Spawn a future and execute it in compact thread pool.

spawn_global

Spawn a future and execute it in global thread pool.

spawn_hb

Spawn a future and execute it in hb thread pool.